+++ doc/changelog.rst
@@ -34,10 +34,11 @@ Added
   gnomonic coordinates. EBSD reference frame documentation.
   (`#204 <https://github.com/pyxem/kikuchipy/pull/204>`_,
   `#215 <https://github.com/pyxem/kikuchipy/pull/215>`_)
-- Geometrical EBSD simulations, projecting a set of Kikuchi bands onto a
-  detector, which can be added to an EBSD signal as markers.
+- Geometrical EBSD simulations, projecting a set of Kikuchi bands and zone axes
+  onto a detector, which can be added to an EBSD signal as markers.
   (`#204 <https://github.com/pyxem/kikuchipy/pull/204>`_,
-  `#219 <https://github.com/pyxem/kikuchipy/pull/219>`_)
+  `#219 <https://github.com/pyxem/kikuchipy/pull/219>`_,
+  `#232 <https://github.com/pyxem/kikuchipy/pull/232>`_)
 - Dependency on the diffsims package (https://github.com/pyxem/diffsims/) for
   handling of electron scattering and diffraction.
   (`#220 <https://github.com/pyxem/kikuchipy/pull/220>`_)